refactor(frontend): extract useComposerPlusMenu hook and guard partialError on loading
