+++ autogpt_platform/frontend/src/app/(platform)/copilot/components/ChatInput/components/ComposerPlusMenu.tsx
@@ -8,7 +8,6 @@ import {
   DropdownMenuTrigger,
 } from "@/components/molecules/DropdownMenu/DropdownMenu";
 import { cn } from "@/lib/utils";
-import { Flag, useGetFlag } from "@/services/feature-flags/use-get-flag";
 import {
   BookOpenIcon,
   CalendarDotsIcon,
@@ -17,8 +16,7 @@ import {
   PlugsConnectedIcon,
   PlusIcon,
 } from "@phosphor-icons/react";
-import { useRef } from "react";
-import { useCopilotModal } from "../../../useCopilotModal";
+import { useComposerPlusMenu } from "./useComposerPlusMenu";
 
 interface Props {
   onFilesSelected: (files: File[]) => void;
@@ -31,22 +29,13 @@ export function ComposerPlusMenu({
   onUseWorkspaceFile,
   disabled,
 }: Props) {
-  const fileInputRef = useRef<HTMLInputElement>(null);
-  const { openModal } = useCopilotModal();
-  const showWorkspaceOption = useGetFlag(Flag.CHAT_WORKSPACE_FILES);
-
-  function openFilePicker() {
-    fileInputRef.current?.click();
-  }
-
-  function handleFileChange(e: React.ChangeEvent<HTMLInputElement>) {
-    const files = Array.from(e.target.files ?? []);
-    if (files.length > 0) {
-      onFilesSelected(files);
-    }
-    // Reset so the same file can be re-selected
-    e.target.value = "";
-  }
+  const {
+    fileInputRef,
+    openModal,
+    showWorkspaceOption,
+    openFilePicker,
+    handleFileChange,
+  } = useComposerPlusMenu({ onFilesSelected });
 
   return (
     <>
